tf: tofu init -migrate-state
Moving the state to local, temporarily, in order to move it to
CloudFlare, to continue my crusade to stop giving Amazon _cents_ in
monthly S3 costs.

Note: this does mean I have created a local terraform.tfstate file
again, which I'm deliberately not committing and not gitignore-ing
because it's very temporary.