deploy on CloudFlare
This deploys both [blog.gary][0] and [beta][1]. The latter is of
particular significance, as it means I could have this served on the
root site, and get rid of all the AWS setup. Out of an abundance of
caution (and because it's late and I'm tired), I'll leave this running
as is overnight, and, if everything seems to go well and I can't think
of an issue with this setup by tomorrow morning, I'll do the transition
then.

[0]: https://blog.gary-verhaegen.workers.dev
[1]: https://beta.cuddly-octo-palm-tree.com