aws: delete s3 bucket
Note that this required me to first delete all of the existing files in
that bucket, per S3 policy.

There was nothing worth keeping in that bucket: I have downloaded all of
the logs (unclear whether I want to keep them, but for now they're on my
machine), and the rest was either easily-recreatable old versions of the
site, or better-deleted sensitive certificate files. See `init.sh` and
the now-deleted `deploy.sh` for details.